Refactoring module card
